Pharmacy Performance Dashboard commelte task 11/6

+++ database/migrations/2025_06_10_080452_add_commission_to_orders_table.php
@@ -19,10 +19,13 @@ return new class extends Migration
     /**
      * Reverse the migrations.
      */
-    public function down(): void
-    {
-        Schema::table('orders', function (Blueprint $table) {
+   public function down(): void
+{
+    Schema::table('orders', function (Blueprint $table) {
+        if (Schema::hasColumn('orders', 'commission')) {
             $table->dropColumn('commission');
-        });
-    }
+        }
+    });
+}
+
 };
